ui: refactor; extract JobDescriptionCell component
Release note: none
